@@ -1,13 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.1"
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "be.createweb.beaconscanner"
-        minSdkVersion 11
-        targetSdkVersion 25
+        minSdkVersion 26
+        targetSdkVersion 29
         versionCode 3
         versionName "1.2"
     }
@@ -40,16 +39,16 @@ android {
 }
 
 dependencies {
-    compile project(':ibeaconscanner') // library to scan for iBeacons
-    compile 'com.android.support:appcompat-v7:25.1.0'
-    compile 'com.android.support:design:25.1.0'
-    compile 'io.reactivex:rxjava:1.2.0' // rxjava
-    compile 'com.tbruyelle.rxpermissions:rxpermissions:0.7.0@aar' // library to handle location permission
-    compile 'com.facebook.stetho:stetho:1.4.1' // stetho to easily see our database
-    debugCompile 'com.squareup.leakcanary:leakcanary-android:1.5' // leak canary to check for leaks
-    releaseCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.5' // leak canary to check for leaks
-    compile 'com.jakewharton:butterknife:8.4.0'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.4.0'
+    implementation project(':ibeaconscanner') // library to scan for iBeacons
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'io.reactivex:rxjava:1.2.0' // rxjava
+    implementation 'com.tbruyelle.rxpermissions:rxpermissions:0.7.0@aar' // library to handle location permission
+    implementation 'com.facebook.stetho:stetho:1.4.1' // stetho to easily see our database
+    debugImplementation 'com.squareup.leakcanary:leakcanary-android:1.5' // leak canary to check for leaks
+    releaseImplementation 'com.squareup.leakcanary:leakcanary-android-no-op:1.5' // leak canary to check for leaks
+    implementation 'com.jakewharton:butterknife:10.2.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.0'
 
-    testCompile 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.12'
 }
