@@ -1,6 +1,6 @@
 package mobi.inthepocket.android.beacons.ibeaconscanner.utils;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.UUID;
 
